client/trailbase-py/pyproject.toml
[tool.poetry]
name = "trailbase"
version = "0.1.0"
description = ""
authors = ["TrailBase <contact@trailbase.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
httpx = "^0.27.2"
pyjwt = "^2.10.0"
cryptography = "^43.0.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
mypy = "^1.13.0"
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.pyright]
venvPath = "."
venv = ".venv"
